Bug fix - Can't drag card, random
Slot types restrictions on cards temporarily disabled in this commit; Fixed a bug in the input_manager.gd that was not searching for the card when clicked on top on. Was colliding with the card slot area, resulting in random bugs of "being able to drag card" and the card being stuck in place. yay