Pre-alpha phase: Functional game logic
With 3 events, it is possible to mantain a high score of "maximum events completed". High score board on the way!

Game logic is still not as I want, of course. Objective now is to fix visual bugs and add minimum player and card interactability, then it can be called alpha.